Refactor SearchPaneView to separate file
